Moving shared iOS+MacOS files to port/apple/.
