+++ xrtl/port/apple/BUILD
@@ -1,0 +1,6 @@
+# Description:
+#  Code common to Apple platforms (iOS, MacOS, etc).
+
+package(default_visibility = ["//xrtl:internal"])
+
+licenses(["notice"])  # Apache 2.0
