+++ xrtl/port/apple/base/BUILD
@@ -1,0 +1,16 @@
+# Description:
+#  Apple port variants of //xrtl/base/ features.
+
+package(default_visibility = ["//xrtl:internal"])
+
+licenses(["notice"])  # Apache 2.0
+
+cc_library(
+    name = "mach_system_clock",
+    srcs = ["mach_system_clock.cc"],
+    hdrs = ["mach_system_clock.h"],
+    deps = [
+        "//xrtl/base:macros",
+        "//xrtl/base:system_clock_hdrs",
+    ],
+)
